% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHM_metrics.R
\name{CHM_metrics}
\alias{CHM_metrics}
\title{CHM Metrics Function}
\usage{
CHM_metrics(chm, agg, gap_thres = 2, metrics = "all", rumple_fast = T)
}
\arguments{
\item{chm}{A canopy height model in SpatRaster format}

\item{agg}{Aggregation factor when calculating the metrics. For instance, if the original resolution is 1m and we want to calculate CHM metrics at 50m grid, agg=50.}

\item{gap_thres}{This is the cutoff height when calculating gap fraction.}

\item{metrics}{Metrics to calculate, defaults to "all" to calculate all metrics. Alternatively metrics could be specified by names: c("chm_maxH","chm_sdH","chm_height_cv","chm_rcv","chm_rms","chm_skewH","chm_kurH","chm_gapFrac","chm_grndFrac","chm_rumple_norm")}

\item{rumple_fast}{Logical. Whether to use a fast approximation of CHM surface area or triangulating point clouds.}
}
\value{
A SpatRaster of metrics, each band being one metric, metric name in band name
}
\description{
This function is to calculate various CHM-based canopy metrics in remote sensing.
}
\details{
We referenced Zhang et al. (2024) when choosing metrics, but also added in some other useful metrics such as RCV.
The normalised rumple index is often the slowest step, so we offer a fast alternative where surface area is calculated from the CHM directly rather than converting the CHM to a triangulated point cloud.
}
