% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHM_rumple_norm.R
\name{CHM_rumple_norm}
\alias{CHM_rumple_norm}
\title{Function to calculate normalised rumple index (rumple_norm)}
\usage{
CHM_rumple_norm(chm_vals, chm_r)
}
\arguments{
\item{chm}{A canopy height model in SpatRaster format}

\item{agg}{Aggregation factor when calculating the metrics. For instance, if the original resolution is 1m and we want to calculate CHM metrics at 50m grid, agg=50.}

\item{metrics}{Metrics to calculate.}
}
\value{
A SpatRaster of metrics, each band being one metric, metric name in band name
}
\description{
Function to calculate normalised rumple index (rumple_norm)
}
