% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_f.R
\name{cov_f}
\alias{cov_f}
\title{LiDAR Metrics Vegetation Cover Helper Function}
\usage{
cov_f(Z, h_cutoff)
}
\value{
Lists to be used in pixel_metrics() and voxel_metrics()
}
\description{
LiDAR Metrics Vegetation Cover Helper Function
}
