% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePAI_f.R
\name{ePAI_f}
\alias{ePAI_f}
\title{LiDAR Metrics ePAI Helper Function}
\usage{
ePAI_f(Z, ReturnNumber, NumberOfReturns, ScanAngleRank, h_cutoff)
}
\value{
Lists to be used in pixel_metrics() and voxel_metrics()
}
\description{
LiDAR Metrics ePAI Helper Function
}
