% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gapFrac_f.R
\name{gapFrac_f}
\alias{gapFrac_f}
\title{LiDAR Metrics Gap Fraction Helper Function}
\usage{
gapFrac_f(Z, gap_thres)
}
\value{
Lists to be used in pixel_metrics() and voxel_metrics()
}
\description{
LiDAR Metrics Gap Fraction Helper Function
}
