% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiDAR_metrics.R
\name{LiDAR_metrics}
\alias{LiDAR_metrics}
\title{LiDAR Metrics Function}
\usage{
LiDAR_metrics(
  las,
  res = 5,
  h_cutoff = 1.3,
  mcc_s = 1.5,
  mcc_t = 0.3,
  cov_grid = 0.25,
  zmax = 35,
  shannon_cut = c(-1, 2, 5, 10, 15, 35),
  vox_res = 0.5,
  L1_range = c(0, 1),
  L2_range = c(1, 10),
  L3_range = c(10, 35)
)
}
\arguments{
\item{las}{A LAS or LAScatalog object}

\item{res}{Resolution in meters}

\item{mcc_s}{For MCC ground classification, see support document for details}

\item{mcc_t}{For MCC ground classification, see support document for details}

\item{cov_grid}{Resolution of grid when tallying pixels when calculating canopy cover}

\item{zmax}{Approximate maximum height of trees}

\item{shannon_cut}{Bins when calculating shannon diversity index for height distribution}

\item{vox_res}{Size of voxels when calculating vegetation volume}

\item{L1_range}{Height boundaries when defining first height layer (same for L2/L3)}
}
\value{
A SpatRaster of metrics, each band being one metric, metric name in band name
}
\description{
This function is used to calculate various LiDAR metrics used in ecology. The function takes LAS or LAScatalog objects created with the lidR package as inputs and returns a multi-layered terra SpatRaster, with each band corresponding to a metric. The choice of metrics, band names, and definitions are mostly based on Shokirov et al. (2023), with some minor changes being made for metrics that were not clearly defined in the paper. Notably, the roughness_L1, roughness_L2, roughness_L3 have not been included in the current edition.
Shokirov, S. et al. (2023) ‘Habitat highs and lows: Using terrestrial and UAV LiDAR for modelling avian species richness and abundance in a restored woodland’, Remote Sensing of Environment. Elsevier, 285, p. 113326. doi: 10.1016/J.RSE.2022.113326.
}
