% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lidar_SHM.R
\name{lidar_SHM}
\alias{lidar_SHM}
\title{SHM creation function}
\usage{
lidar_SHM(
  las,
  DTM_fp,
  DTM_res = 2.3,
  max_shrub_ht = 1.025,
  stretch_factor = 22,
  mcc_s = 0.5,
  mcc_t = 0.525,
  shm_res = 0.2
)
}
\arguments{
\item{las}{A LAS or LAScatalog object created by the \code{lidR} package}

\item{DTM_fp}{Character. Filepath to a standard DTM in the format "D:/path/DTM.tif". This is used to filter out trees before stretching. The quality of this DTM is not critical as the normalisation is later reversed, but a raster of reasonable resolution (e.g. 2m) is recommended.}

\item{DTM_res}{Numeric. Resolution of the new DTM produced for the second ground classification step. Defaults to 2.3.}

\item{max_shrub_ht}{Numeric. Points above this height is regarded as trees and filtered out after first normalisation. Defaults to 1.025.}

\item{stretch_factor}{Numeric. Determines how much the point cloud is being stretched out before ground classfication. Defaults to 22.}

\item{mcc_s}{Numeric. Scale parameter fed into the \code{mcc()} function during ground classification. Defaults to 0.5.}

\item{mcc_t}{Numeric. Curvature threshold fed into the \code{mcc()} function during ground classification. Defaults to 0.525.}

\item{shm_res}{Numeric. Resolution of the output SHM. Defaults to 0.2m.}
}
\value{
A SpatRaster. The shrub height model (SHM).
}
\description{
A function to create shrub height models (SHM) from LiDAR point clouds. The function takes a LAS or LAScatalog object and builds an SHM based on the parameters specified.
}
\details{
Default parameter values are from Chan et al. (2026) and represent values optimised for a airborne LiDAR scan in Scotland. Ericaceous shrubs in Scotland are short (mean height 0.36m), so some adjustments might need to be made if the study system contains taller shrubs.
}
