% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{q99_f}
\alias{q99_f}
\title{LiDAR Metrics Helper Functions}
\usage{
q99_f(z, zmax)
}
\value{
Lists to be used in pixel_metrics() and voxel_metrics()
}
\description{
LiDAR Metrics Helper Functions
}
