% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roughness_metrics_f.R
\name{roughness_metrics_f}
\alias{roughness_metrics_f}
\title{LiDAR Metrics Roughness Helper Function}
\usage{
roughness_metrics_f(Z, shannon_cut)
}
\value{
Lists to be used in pixel_metrics() and voxel_metrics()
}
\description{
LiDAR Metrics Roughness Helper Function
}
