% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vox_f.R
\name{vox_f}
\alias{vox_f}
\title{LiDAR Metrics Voxel Helper Function}
\usage{
vox_f(vox_res)
}
\value{
Lists to be used in pixel_metrics() and voxel_metrics()
}
\description{
LiDAR Metrics Voxel Helper Function
}
