% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LiDAR_metrics.R
\name{LiDAR_metrics}
\alias{LiDAR_metrics}
\title{LiDAR Metrics Function}
\usage{
LiDAR_metrics(
  las,
  ground_classified = F,
  res = 5,
  h_cutoff = 1.3,
  gap_thres = 2,
  mcc_s = 1.5,
  mcc_t = 0.3,
  zmax = 35,
  shannon_cut = c(-1, 2, 5, 10, 15, 35),
  vox_res = 0.5,
  L1_range = c(0, 1),
  L2_range = c(1, 10),
  L3_range = c(10, 35),
  metrics = "all"
)
}
\arguments{
\item{las}{A LAS or LAScatalog object}

\item{ground_classified}{True if the LAS object supplied is already ground classified}

\item{res}{Resolution in meters}

\item{h_cutoff}{Ground and shrub points are removed when calculating metrics such as SD, skew and kurtosis. This represents the cutoff height, 1.3m by default (see Shokirov et al. (2023)).}

\item{gap_thres}{Threshold for calculating gap fraction. Default at 2m.}

\item{mcc_s}{For MCC ground classification, see support document for details}

\item{mcc_t}{For MCC ground classification, see support document for details}

\item{zmax}{Approximate maximum height of trees}

\item{shannon_cut}{Bins when calculating shannon diversity index for height distribution}

\item{vox_res}{Size of voxels when calculating vegetation volume}

\item{L1_range}{Height boundaries when defining first height layer (same for L2/L3)}

\item{metrics}{Metrics to calculate. Defaults to "all", but could be any subset of c("lidar_maxH", "lidar_meanH", "lidar_stdH", "lidar_skewH", "lidar_kurH", "lidar_p_05", "lidar_p_10", "lidar_p_25", "lidar_p_50", "lidar_p_75", "lidar_p_90", "lidar_p_95", "lidar_p_999", "lidar_vci_2m", "lidar_vci_5m", "lidar_vci_10m", "lidar_vci_15m", "lidar_vci_20m", "lidar_Cov","lidar_gapFrac", "lidar_grndFrac", "lidar_height_cv","lidar_rcv","lidar_rms", "lidar_canopy_shannon", "lidar_Tvolume", "lidar_vlayer_L1", "lidar_vlayer_L2", "lidar_vlayer_L3", "lidar_meanH_L1", "lidar_sdH_L1", "lidar_meanH_L2", "lidar_sdH_L2", "lidar_meanH_L3", "lidar_sdH_L3", "lidar_vci_L1", "lidar_vci_L2", "lidar_vci_L3")}

\item{cov_grid}{Resolution of grid when tallying pixels when calculating canopy cover}
}
\value{
A SpatRaster of metrics, each band being one metric, metric name in band name
}
\description{
This function is used to calculate various LiDAR metrics used in ecology. The function takes LAS or LAScatalog objects created with the lidR package as inputs and returns a multi-layered terra SpatRaster, with each band corresponding to a metric.
}
